Fix issue with default allocator lifetime
A released buffer pointer can live longer than the FlatBufferBuilder
and hence longer than the default allocator. This fix gives the default
allocator static storage.

Before the previous custom allocators fix, the tests were only working
because std::bind was copying the allocator by value. This worked
because the default allocator is trivial, but it is a bug if the
allocator is stateful. The existing code stores custom allocators by
reference anyway, so it is expected that the caller should be aware of
the allocator's lifetime.

Now if the caller does not provide an allocator, things "just work", and
if they do provide one, it is expected to survive any released buffer
pointers. (This makes sense because the allocator is the one doing the
releasing!)

Note that the static storage could be eliminated entirely by forcing the
caller to always provide an allocator, but I did not want to change the
existing interface.